# perf(db): open MDBX DBIs only once at startup