Part 2 of #18292: Cache DBIs at startup time to reuse throughout the program's lifetime instead of just per transaction.

<img width="1267" height="362" alt="image" src="https://github.com/user-attachments/assets/472005a2-1095-42bb-9739-edaba883a8f1" />

<img width="836" height="149" alt="image" src="https://github.com/user-attachments/assets/5efc8e9b-f0a1-4dc1-bef2-466fabe9ec30" />

---

`reth-bench` for the first 300 blocks of Shanghai:

```
$ reth-bench new-payload-fcu --advance 300
```

f66e19717135db253462069991f41950201720f6 (`main`):
```
2025-09-13T09:46:50.043037Z  INFO Total Ggas/s: 0.6499 total_duration=7.062721087s total_gas_used=4590139398 blocks_processed=300
```

This branch on f66e19717135db253462069991f41950201720f6:
```
2025-09-13T09:56:17.725214Z  INFO Total Ggas/s: 0.6967 total_duration=6.588003416s total_gas_used=4590139398 blocks_processed=300
```

Speedup:
- Ggas/s: 0.6499 -> 0.6967 ~ 7.2% higher execution throughput
- Duration: 7.062721087s -> 6.588003416s ~ 7.2% lower payload latency
- Best win: From 39.2ms down to 13.3ms payload latency